Remove java-repo-tools directory.
It is not needed anymore, now that we package the parent pom in Maven
central.

Rather than using submodules, this time I clone the java-repo-tools repo
in the test script. (I probably should have done that for the other
repos, too, but it felt especially important for this quickstart repo).